Moved Java files to default Maven structure
