+++ Spring-Hibernate-Train/src/main/java/be/kapture/training/dao/PersonDAOImpl.java
@@ -1,0 +1,38 @@
+package be.kapture.training.dao;
+
+import java.util.List;
+
+import org.hibernate.Session;
+import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
+
+import be.kapture.training.model.*;
+
+/**
+ * Created by vanmoj1 on 3/04/2017.
+ */
+public class PersonDAOImpl implements PersonDAO {
+    private SessionFactory sessionFactory;
+
+    public void setSessionFactory(SessionFactory sessionFactory) {
+        this.sessionFactory = sessionFactory;
+    }
+
+    @Override
+    public void save(Person p) {
+        Session session = this.sessionFactory.openSession();
+        Transaction tx = session.beginTransaction();
+        session.persist(p);
+        tx.commit();
+        session.close();
+    }
+
+    @SuppressWarnings("unchecked")
+    @Override
+    public List<Person> list() {
+        Session session = this.sessionFactory.openSession();
+        List<Person> personList = session.createQuery("from Person").list();
+        session.close();
+        return personList;
+    }
+}
