+++ Spring-Hibernate-Train/src/main/java/be/kapture/training/model/Customer.java
