+++ Spring-Hibernate-Train/src/main/java/be/kapture/training/model/Person.java
@@ -1,0 +1,58 @@
+package be.kapture.training.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+/**
+ * Created by vanmoj1 on 3/04/2017.
+ */
+
+/**
+ * Entity bean with JPA annotations
+ * Hibernate provides JPA implementation
+ */
+@Entity
+@Table(name = "Person")
+public class Person {
+    @Id
+    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private String name;
+
+    private String country;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getCountry() {
+        return country;
+    }
+
+    public void setCountry(String country) {
+        this.country = country;
+    }
+
+    @Override
+    public String toString() {
+        return "id=" + id + ", name=" + name + ", country=" + country;
+    }
+}
