+++ Spring-Hibernate-Train/src/test/java/Slf4jLoggerTest.java
